moved notifications test to the test class
